refactor: RecruitingDetail 1차 분리
