+++ src/components/common/Pagination.jsx
@@ -1,0 +1,71 @@
+import React from 'react';
+import { ChevronLeft, ChevronRight } from 'lucide-react';
+import './Pagination.css';
+
+const Pagination = ({ 
+  currentPage, 
+  totalPages, 
+  startIndex, 
+  endIndex, 
+  totalItems,
+  onPageChange 
+}) => {
+  const handlePrevious = () => {
+    onPageChange(Math.max(currentPage - 1, 1));
+  };
+
+  const handleNext = () => {
+    onPageChange(Math.min(currentPage + 1, totalPages));
+  };
+
+  const handlePageClick = (pageNum) => {
+    onPageChange(pageNum);
+  };
+
+  if (totalPages <= 1) {
+    return null;
+  }
+
+  return (
+    <div className="pagination">
+      <div className="pagination-info">
+        <span>
+          {startIndex + 1}-{Math.min(endIndex, totalItems)} / {totalItems}개
+        </span>
+      </div>
+      <div className="pagination-controls">
+        <button
+          className="pagination-btn"
+          onClick={handlePrevious}
+          disabled={currentPage === 1}
+        >
+          <ChevronLeft size={16} />
+          이전
+        </button>
+        
+        <div className="pagination-numbers">
+          {Array.from({ length: totalPages }, (_, i) => i + 1).map(pageNum => (
+            <button
+              key={pageNum}
+              className={`pagination-number ${pageNum === currentPage ? 'active' : ''}`}
+              onClick={() => handlePageClick(pageNum)}
+            >
+              {pageNum}
+            </button>
+          ))}
+        </div>
+        
+        <button
+          className="pagination-btn"
+          onClick={handleNext}
+          disabled={currentPage === totalPages}
+        >
+          다음
+          <ChevronRight size={16} />
+        </button>
+      </div>
+    </div>
+  );
+};
+
+export default Pagination;
