+++ src/components/recruiting/BulkStatusChangeModal.css
@@ -1,0 +1,241 @@
+/* 일괄 상태 변경 모달 스타일 */
+.bulk-change-modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  backdrop-filter: blur(4px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.bulk-change-modal-container {
+  background: linear-gradient(135deg, var(--color-gray-8) 0%, var(--color-gray-7) 100%);
+  border: 1px solid var(--color-gray-6);
+  border-radius: 16px;
+  box-shadow: 0 25px 50px -12px rgba(0, 0, 0, 0.8);
+  max-width: 500px;
+  width: 90%;
+  max-height: 80vh;
+  overflow: hidden;
+}
+
+.bulk-change-modal-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1.5rem 2rem;
+  background: linear-gradient(135deg, var(--color-green-0) 0%, #22c55e 100%);
+  position: relative;
+  overflow: hidden;
+}
+
+.bulk-change-modal-title {
+  color: var(--color-white);
+  font-size: 1.25rem;
+  font-weight: 700;
+  margin: 0;
+  text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
+}
+
+.bulk-change-modal-close-btn {
+  background: none;
+  border: none;
+  color: var(--color-white);
+  font-size: 1.5rem;
+  cursor: pointer;
+  width: 32px;
+  height: 32px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 50%;
+  transition: all 0.3s;
+}
+
+.bulk-change-modal-close-btn:hover:not(:disabled) {
+  background-color: rgba(255, 255, 255, 0.2);
+}
+
+.bulk-change-modal-close-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.bulk-change-modal-content {
+  padding: 2rem;
+}
+
+.bulk-change-description {
+  color: var(--color-gray-2);
+  font-size: 0.875rem;
+  margin-bottom: 2rem;
+  text-align: center;
+  line-height: 1.5;
+}
+
+.bulk-change-controls {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.bulk-count-section {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.bulk-count-label {
+  color: var(--color-white);
+  font-size: 0.875rem;
+  font-weight: 600;
+}
+
+.bulk-count-input-wrapper {
+  position: relative;
+  width: 120px;
+}
+
+.bulk-count-input {
+  width: 100%;
+  padding: 0.75rem 2rem 0.75rem 1rem;
+  background-color: var(--color-gray-8);
+  border: 1px solid var(--color-gray-5);
+  border-radius: 8px;
+  color: var(--color-white);
+  font-size: 1rem;
+  font-weight: 500;
+  text-align: center;
+}
+
+.bulk-count-input:focus {
+  outline: none;
+  border-color: var(--color-green-0);
+  box-shadow: 0 0 0 3px rgba(34, 197, 94, 0.1);
+}
+
+.bulk-count-input:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.bulk-count-suffix {
+  position: absolute;
+  right: 0.75rem;
+  top: 50%;
+  transform: translateY(-50%);
+  color: var(--color-gray-3);
+  font-size: 0.875rem;
+  font-weight: 500;
+  pointer-events: none;
+}
+
+.bulk-status-section {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.bulk-status-label {
+  color: var(--color-white);
+  font-size: 0.875rem;
+  font-weight: 600;
+}
+
+.bulk-status-buttons {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 1rem;
+}
+
+.bulk-status-btn {
+  padding: 1rem 1.5rem;
+  border: 2px solid;
+  border-radius: 8px;
+  font-size: 0.875rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.3s;
+  text-align: center;
+}
+
+.bulk-status-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.bulk-status-btn.first-pass {
+  background-color: transparent;
+  border-color: #3b82f6;
+  color: #3b82f6;
+}
+
+.bulk-status-btn.first-pass:hover:not(:disabled) {
+  background-color: #3b82f6;
+  color: var(--color-white);
+}
+
+.bulk-status-btn.final-pass {
+  background-color: transparent;
+  border-color: var(--color-green-0);
+  color: var(--color-green-0);
+}
+
+.bulk-status-btn.final-pass:hover:not(:disabled) {
+  background-color: var(--color-green-0);
+  color: var(--color-white);
+}
+
+.bulk-status-btn.failed {
+  background-color: transparent;
+  border-color: #ef4444;
+  color: #ef4444;
+}
+
+.bulk-status-btn.failed:hover:not(:disabled) {
+  background-color: #ef4444;
+  color: var(--color-white);
+}
+
+.bulk-status-btn.pending {
+  background-color: transparent;
+  border-color: #f59e0b;
+  color: #f59e0b;
+}
+
+.bulk-status-btn.pending:hover:not(:disabled) {
+  background-color: #f59e0b;
+  color: var(--color-white);
+}
+
+.bulk-change-loading {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+  margin-top: 1.5rem;
+  padding: 1rem;
+  background-color: var(--color-gray-6);
+  border-radius: 8px;
+  color: var(--color-white);
+  font-size: 0.875rem;
+}
+
+.loading-spinner {
+  width: 20px;
+  height: 20px;
+  border: 2px solid var(--color-gray-5);
+  border-top: 2px solid var(--color-green-0);
+  border-radius: 50%;
+  animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
